Реорганизовал h_cycler.cpp
