+++ server/entities/cycler.cpp
@@ -1,0 +1,152 @@
+/***
+*
+*	Copyright (c) 1996-2002, Valve LLC. All rights reserved.
+*	
+*	This product contains software technology licensed from Id 
+*	Software, Inc. ("Id Technology").  Id Technology (c) 1996 Id Software, Inc. 
+*	All Rights Reserved.
+*
+*   Use, distribution, and modification of this source code and/or resulting
+*   object code is restricted to non-commercial enhancements to products from
+*   Valve LLC.  All other use, distribution, or modification is prohibited
+*   without written permission from Valve LLC.
+*
+****/
+/*
+
+===== h_cycler.cpp ========================================================
+
+  The Halflife Cycler Monsters
+
+*/
+
+#include "cycler.h"
+
+#define TEMP_FOR_SCREEN_SHOTS
+#ifdef TEMP_FOR_SCREEN_SHOTS //===================================================
+
+BEGIN_DATADESC( CCycler )
+	DEFINE_FIELD( m_animate, FIELD_INTEGER ),
+END_DATADESC()
+
+LINK_ENTITY_TO_CLASS( cycler, CGenericCycler );
+
+// Cycler member functions
+
+void CCycler :: GenericCyclerSpawn(char *szModel, Vector vecMin, Vector vecMax)
+{
+	if (!szModel || !*szModel)
+	{
+		ALERT(at_error, "cycler at %.0f %.0f %0.f missing modelname", GetAbsOrigin().x, GetAbsOrigin().y, GetAbsOrigin().z );
+		REMOVE_ENTITY(ENT(pev));
+		return;
+	}
+
+	pev->classname		= MAKE_STRING("cycler");
+	PRECACHE_MODEL( szModel );
+	SET_MODEL(ENT(pev),	szModel);
+
+	CCycler::Spawn( );
+
+	UTIL_SetSize(pev, vecMin, vecMax);
+}
+
+
+void CCycler :: Spawn( )
+{
+	InitBoneControllers();
+	pev->solid			= SOLID_SLIDEBOX;
+	pev->movetype		= MOVETYPE_NONE;
+	pev->takedamage		= DAMAGE_YES;
+	pev->effects		= 0;
+	pev->health			= 80000;// no cycler should die
+	pev->yaw_speed		= 5;
+	pev->ideal_yaw		= GetAbsAngles().y;
+	ChangeYaw( 360 );
+	
+	m_flFrameRate		= 75;
+	m_flGroundSpeed		= 0;
+
+	pev->nextthink		+= 1.0;
+
+	ResetSequenceInfo( );
+
+	if (pev->sequence != 0 || pev->frame != 0)
+	{
+		m_animate = 0;
+		pev->framerate = 0;
+	}
+	else
+	{
+		m_animate = 1;
+	}
+}
+
+//
+// cycler think
+//
+void CCycler :: Think( void )
+{
+	pev->nextthink = gpGlobals->time + 0.1;
+
+	if (m_animate)
+	{
+		StudioFrameAdvance ( );
+	}
+	if (m_fSequenceFinished && !m_fSequenceLoops)
+	{
+		// ResetSequenceInfo();
+		// hack to avoid reloading model every frame
+		pev->animtime = gpGlobals->time;
+		pev->framerate = 1.0;
+		m_fSequenceFinished = FALSE;
+		m_flLastEventCheck = gpGlobals->time;
+		pev->frame = 0;
+		if (!m_animate)
+			pev->framerate = 0.0;	// FIX: don't reset framerate
+	}
+}
+
+//
+// CyclerUse - starts a rotation trend
+//
+void CCycler :: Use ( CBaseEntity *pActivator, CBaseEntity *pCaller, USE_TYPE useType, float value )
+{
+	m_animate = !m_animate;
+	if (m_animate)
+		pev->framerate = 1.0;
+	else
+		pev->framerate = 0.0;
+}
+
+//
+// CyclerPain , changes sequences when shot
+//
+//void CCycler :: Pain( float flDamage )
+int CCycler :: TakeDamage( entvars_t* pevInflictor, entvars_t* pevAttacker, float flDamage, int bitsDamageType )
+{
+	if (m_animate)
+	{
+		pev->sequence++;
+
+		ResetSequenceInfo( );
+
+		if (m_flFrameRate == 0.0)
+		{
+			pev->sequence = 0;
+			ResetSequenceInfo( );
+		}
+		pev->frame = 0;
+	}
+	else
+	{
+		pev->framerate = 1.0;
+		StudioFrameAdvance ( 0.1 );
+		pev->framerate = 0;
+		ALERT( at_console, "sequence: %d, frame %.0f\n", pev->sequence, pev->frame );
+	}
+
+	return 0;
+}
+
+#endif
