+++ server/entities/cycler_sprite.h
@@ -1,0 +1,51 @@
+/***
+*
+*	Copyright (c) 1996-2002, Valve LLC. All rights reserved.
+*	
+*	This product contains software technology licensed from Id 
+*	Software, Inc. ("Id Technology").  Id Technology (c) 1996 Id Software, Inc. 
+*	All Rights Reserved.
+*
+*   Use, distribution, and modification of this source code and/or resulting
+*   object code is restricted to non-commercial enhancements to products from
+*   Valve LLC.  All other use, distribution, or modification is prohibited
+*   without written permission from Valve LLC.
+*
+****/
+/*
+
+===== h_cycler.cpp ========================================================
+
+  The Halflife Cycler Monsters
+
+*/
+
+#pragma once
+
+#include "extdll.h"
+#include "util.h"
+#include "cbase.h"
+#include "monsters.h"
+#include "animation.h"
+#include "weapons.h"
+#include "player.h"
+
+class CCyclerSprite : public CBaseEntity
+{
+	DECLARE_CLASS( CCyclerSprite, CBaseEntity );
+public:
+	void Spawn( void );
+	void Think( void );
+	void Use( CBaseEntity *pActivator, CBaseEntity *pCaller, USE_TYPE useType, float value );
+	virtual int ObjectCaps( void ) { return (BaseClass :: ObjectCaps() | FCAP_IMPULSE_USE); }
+	virtual int TakeDamage( entvars_t* pevInflictor, entvars_t* pevAttacker, float flDamage, int bitsDamageType );
+	void Animate( float frames );
+
+	DECLARE_DATADESC();
+
+	inline int ShouldAnimate( void ) { return m_animate && m_maxFrame > 1.0; }
+
+	int	m_animate;
+	float	m_lastTime;
+	float	m_maxFrame;
+};
