+++ server/entities/cycler_wreckage.h
@@ -1,0 +1,44 @@
+/***
+*
+*	Copyright (c) 1996-2002, Valve LLC. All rights reserved.
+*	
+*	This product contains software technology licensed from Id 
+*	Software, Inc. ("Id Technology").  Id Technology (c) 1996 Id Software, Inc. 
+*	All Rights Reserved.
+*
+*   Use, distribution, and modification of this source code and/or resulting
+*   object code is restricted to non-commercial enhancements to products from
+*   Valve LLC.  All other use, distribution, or modification is prohibited
+*   without written permission from Valve LLC.
+*
+****/
+/*
+
+===== h_cycler.cpp ========================================================
+
+  The Halflife Cycler Monsters
+
+*/
+
+#pragma once
+
+#include "extdll.h"
+#include "util.h"
+#include "cbase.h"
+#include "monsters.h"
+#include "animation.h"
+#include "weapons.h"
+#include "player.h"
+
+class CWreckage : public CBaseMonster
+{
+	DECLARE_CLASS( CWreckage, CBaseMonster );
+
+	DECLARE_DATADESC();
+
+	void Spawn( void );
+	void Precache( void );
+	void Think( void );
+
+	int m_flStartTime;
+};
